requirements.txt
asgiref==3.11.1
Django==6.0.2
sqlparse==0.5.5
djangorestframework
